♻️Refactor: 직원, 관리자 처음회원가입 설정 페이지 리팩토링 완료
직원, 관리자 처음회원가입 설정 페이지 리팩토링 완료
