+++ src/Components/common/AppTitle.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const AppTitle = () => {
+  return (
+    <h2 className="scroll-m-20 text-2xl font-bold tracking-tight text-center">Attendance-App</h2>
+  );
+};
+
+export default AppTitle;
