+++ src/constant/steps.js
@@ -1,3 +1,7 @@
-const steps = [];
-
-export default steps;
+export const COMPANY_SETTING_STEPS = [
+  "회사 기본 설정",
+  "회사 직무 설정",
+  "야간 공휴일 설정",
+  "근무지 추가",
+  "직원 초대",
+];
