+++ src/hooks/useSettingManager.ts
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const useSettingManager = () => {};
